news_generator/requirements.txt
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.2.3
django-cors-headers==3.7.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.5.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.3
mysqlclient==2.0.3
newsapi-python==0.2.6
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.61.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.10.0.0
urllib3==1.26.5
w3lib==1.22.0
zipp==3.4.1
zope.interface==5.4.0
